Changes before new site master table

+++ machine.py
@@ -6,17 +6,20 @@ from .models import ClientMachine
 machine = Blueprint('machine', __name__)
 
 
-@machine.route('/machines')
-@login_required
-def machines():
-    if not current_user.is_admin():
-        return render_template('page_403.html')
-    return render_template('machines.html')
+# @machine.route('/machines')
+# @login_required
+# def machines():
+#     if not current_user.is_root():
+#         return render_template('page_403.html')
+#     return render_template('machines.html')
+# <li><a class="dropdown-item" href="{{ url_for('machine.machines') }}">Machines</a></li>
 
 
 @machine.route('/getMachine/<id>')
 @login_required
 def get_machine_one(id):
+    if not current_user.is_root():
+        return jsonify({'error': 'Only root admins can access this section.'})
     machine = ClientMachine.get_machine(id)
     print(machine)
     return jsonify(machine)
@@ -25,6 +28,8 @@ def get_machine_one(id):
 @machine.route('/getMachine')
 @login_required
 def get_machine():
+    if not current_user.is_root():
+        return jsonify({'error': 'Only root admins can access this section.'})
     machines = ClientMachine.get_machines()
     print(machines)
     return jsonify({'machines': machines})
@@ -33,6 +38,8 @@ def get_machine():
 @machine.route('/updateMachine/<id>', methods=['PUT'])
 @login_required
 def update_machine(id):
+    if not current_user.is_root():
+        return jsonify({'error': 'Only root admins can access this section.'})
     machineData = request.form
     machine = ClientMachine.get_machine_by_address_not_id(
         machineData['address'], id)
@@ -41,13 +48,16 @@ def update_machine(id):
         return jsonify({'error': 'A machine with the same address already exists.'})
 
     result = ClientMachine.update_machine(ClientMachine(id=id, address=machineData['address'],
-                                                        description=machineData['description']))
+                                                        description=machineData['description'],
+                                                        van_number=machineData['van_number']))
     return jsonify(result)
 
 
 @machine.route('/addMachine', methods=['POST'])
 @login_required
 def add_machine():
+    if not current_user.is_root():
+        return jsonify({'error': 'Only root admins can access this section.'})
     machineData = request.form
     print(machineData)
 
@@ -57,12 +67,14 @@ def add_machine():
         return jsonify({'error': 'A machine with the same address already exists.'})
 
     result = ClientMachine.create_machine(
-        machineData['address'], machineData['description'])
+        machineData['address'], machineData['description'], machineData['van_number'])
     return jsonify(result)
 
 
 @machine.route('/deleteMachine/<id>', methods=['DELETE'])
 @login_required
 def delete_machine(id):
+    if not current_user.is_root():
+        return jsonify({'error': 'Only root admins can access this section.'})
     result = ClientMachine.delete_machine(id)
     return jsonify(result)
